Allow posts to live in both hemispheres.
I wanted my week-in-review posts to be available on both sides, with
different layout and different URLs. This required setting up two
filesystem sources, since each source has only one route. And that
led to further complications. But now it can be done, by making one
a symlink to the other.