Slight readability improvements.
The pink background on the blog page is a problem, but if I tone it
down, the home page loses its vibrance and makes me sad. A possible
solution is to change the inner pages, but change them enough to feel
like a choice, not cowardice. Meanwhile, lightening the text helps a
little.